feat: moved patch recon loss to loss module
